Use per-server cache
This likely will fix #47. It fixes CN server using global game data cache
triggered by the following:

 1. Open the launcher in global mode
 2. The launcher caches the global data
 3. Use the launcher to locate the launcher folder, update the config to point
    to the CN server
 4. Restart the launcher, now the launcher recognizes it's using CN server, but
    still downloads the global game due to the cache.

This patch adds server as the suffix of the cache key of the game data and
patch, so that it won't use the cached data for the wrong server.

Also, avoid downloading the data if it's already cached.